Refactor DCT code out of too big class
