+++ source/lib/analyzer/DCTTransformsNative.h
